+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - added function `Pdf::with_setname_and_nmem`
+- changed return type of `PdfSet::uncertainty` to `Result<_>`
 
 ## [0.1.10] - 18/03/2021